fix test 1 failing

+++ Term 1/Lesson 13 - 35/Lesson_34_Activity_Two.java
@@ -19,7 +19,7 @@ class Lesson_34_Activity_Two {
 
     public static double average(int[] a)
     {
-        int out = 0;
+        double out = 0;
         for (int num : a)
         {
             out += num;
